Fix more warnings from gcc and lint.
These should allow for a clean build in the NetBSD tree with WARNS=4.
Thanks to Alistair Crooks for the patches!

darcs-hash:20090522090853-ced27-5f1132e73422c5592d1c40612b4e6283e5d09916.gz

+++ lib/regerror.c
@@ -62,8 +62,10 @@ regerror(int errcode, const regex_t *preg, char *errbuf, size_t errbuf_size)
   const char *err;
   size_t err_len;
 
+  /*LINTED*/(void)&preg;
   if (errcode >= 0
-      && errcode < (sizeof(tre_error_messages) / sizeof(*tre_error_messages)))
+      && errcode < (int)(sizeof(tre_error_messages)
+			 / sizeof(*tre_error_messages)))
     err = gettext(tre_error_messages[errcode]);
   else
     err = gettext("Unknown error");
